Use proper folder structure for pro (#2104)
